+++ src/__test__/utils/editOntology/setOntologyUpdateNode.test.js
@@ -42,7 +42,7 @@ describe('setOntologyUpdateNode', () => {
     httpCall.mockImplementationOnce(() => ({ error: true }))
 
     store.getState = jest.fn().mockImplementation(() => ({
-      stylingNodeCaptionProperty: 'rdfsLabel',
+      userDefinedNodeStyling: { stylingNodeCaptionProperty: 'rdfsLabel' },
       classesFromApi,
       updatedNodes: [],
     }))
@@ -66,7 +66,7 @@ describe('setOntologyUpdateNode', () => {
     httpCall.mockImplementationOnce(() => ({ data: {} }))
 
     store.getState = jest.fn().mockImplementation(() => ({
-      stylingNodeCaptionProperty: 'rdfsLabel',
+      userDefinedNodeStyling: { stylingNodeCaptionProperty: 'rdfsLabel' },
       classesFromApi,
       updatedNodes: [],
     }))
@@ -97,7 +97,7 @@ describe('setOntologyUpdateNode', () => {
     }))
 
     store.getState = jest.fn().mockImplementationOnce(() => ({
-      stylingNodeCaptionProperty: 'rdfsLabel',
+      userDefinedNodeStyling: { stylingNodeCaptionProperty: 'rdfsLabel' },
       classesFromApi,
       updatedNodes: [],
     }))
@@ -112,7 +112,7 @@ describe('setOntologyUpdateNode', () => {
     expect(updateNodes).toHaveBeenLastCalledWith(
       {
         id: '1',
-        label: 'New node',
+        label: 'New\nnode',
         rdfsLabel: 'New node',
         'Business Area': 'Communications',
         nodeId: 1,
@@ -123,10 +123,13 @@ describe('setOntologyUpdateNode', () => {
       }
     )
 
+    const classesFromApiOutput = JSON.parse(JSON.stringify(newClassesFromApi))
+    classesFromApiOutput['1'].label = 'New\nnode'
+
     expect(setStoreState.mock.calls).toEqual([
       [
         'classesFromApi',
-        newClassesFromApi
+        classesFromApiOutput
       ],
       [
         'updatedNodes',