+++ src/utils/freeTextSearch/clearElement.js
@@ -8,30 +8,38 @@ import updateNodes from '../nodesEdgesUtils/updateNodes'
  */
 const clearElement = () => {
   const {
-    stylingNodeBackgroundColor,
-    stylingEdgeLineColor,
+    userDefinedNodeStyling,
+    globalNodeStyling,
+    userDefinedEdgeStyling,
+    globalEdgeStyling,
     freeTextPrevSelectedElement,
     freeTextSelection
   } = store.getState()
 
   if (!freeTextPrevSelectedElement) return false
 
-  const { id } = freeTextPrevSelectedElement
+  const { id, userDefined } = freeTextPrevSelectedElement
   const color = freeTextPrevSelectedElement.color || {}
+  const width = freeTextPrevSelectedElement.width || 1
 
-  const type = freeTextSelection[id]
+  if (!freeTextSelection[id]) return false
+
+  const { type } = freeTextSelection[id]
 
   if (type === 'edge') {
+    const { stylingEdgeLineColor } = userDefined ? userDefinedEdgeStyling : globalEdgeStyling
     color.color = stylingEdgeLineColor
 
     return updateEdges(
       {
         id,
         color,
-        width: 1
+        width
       }
     )
   }
+
+  const { stylingNodeBackgroundColor } = userDefined ? userDefinedNodeStyling : globalNodeStyling
 
   color.background = stylingNodeBackgroundColor